Update SpreadsheetReader.py
Added function to read in covering dates from CSV file for use with the pieces.
Having added that, added code to compare the opening dates that were generated from extracting the dates from the summary description and the supplied covering dates. It now flags in the summary document when there are discrepancies.