Switch to using extracted dates other covering dates when different & test for deceased in notes
Added option to feed covering date into extracting date from the descriptive column so that it could be used as the default date when no date was found.
Since the extracted date should be used by preference when it is different to the covering date, the extracted dates (with the covering date for when no dates could be identified) is now the preferred list for generating the opening dates.
A number of parameter names have been updated to clarify.
An option to not print out the full details in the summery has been added as a default. This reduces the output to problems rather than problems and notifications. Sending the 'full' parameter as true will print out both.
A further test has been added in to check the deceased redactions.